added timings
